requirements.txt
pandas
plotly
Pillow
scikit_learn
streamlit
streamlit_option_menu
numpy
joblib
pillow